refactor cljs-flux usage (see notes)
this commit refactors the flux event handling to use a single
dispatcher instead of a separate for each type of event.

the handling of different event types is now handled via an
easily extendable multimethod.

one potential problem with this implementation could be that
cljs-flux has some sort of lock on dispatchers that prevents
them from dispatching events from inside event handlers.

the clear-loading event is this kind of thing but I believe
it works because a timeout is used, decoupling it from the actual
handler.

if this turns into a problem it will be easy to swap out cljs-flux
with something that acts more like a queue.