Move reaction and comment add and delete defmethods
Including the action creators for reaction and comment into the ws-interaction-client was causing a circular dependency since this is needed also for AP in the activity actions, but comment and reaction needs the activity actions to refresh the post data. To fix it this moves the client defmethod declaration into the action creators so they are called directly from within it.