User management refactor
Cleanup the logic behind user invite, user type switch and admin and author add and remove. Part of the logic was in oc.web.api and part in oc.web.actions, now the full logic is moved to the team actions namespace and only the primitive API calls are left in the api namespace. Use shared callback for the api calls to avoid code duplication and missed refreshes when data change.