Logic behind the invite people tooltip
When dismissing the add post tooltip set the cookie for the invite people tooltip and show it. Show the invite people tooltip when there is only one user on the team and the user can invite people, ie the link is present. Remove the cookie when one of these is not true.