Add components for special 404 cases
Adds 2 new components called actibity-not-found and activity-rmeoved to be used in place of our usual 404 page. The first is used when the user is not logged in and the org/section/activity is not found and prompts to login to see the content, the second is used when the user is logged in but the section/activity required are not found. The special case where the user is actually logged in and the org is not found is handled by showing the usual 404.