Refresh and reconnect if JWT is expired
Refactor a bit the JWT dispatch and update to make them available also to the ws clients. Change the ws client dir structure to add a utility namespace. Check if the jwt is expired before connecting and check again after if jwt is refiused, report to sentry when jwt is not valid but not expired.